Handle Kibana password too
